FixBug : fixed where some stimul report values were not passing properly , fixed where order by lists was put after the pagination functions , fixed where int.parse was throwing exception when putting a big chunk of digits

+++ FreightAccounting.WPF/AddDebtorWindow.xaml.cs
@@ -119,7 +119,7 @@ public partial class AddDebtorWindow : Window
         }
 
 
-        var isNumeric = int.TryParse(txtDebtAmount.Text.Replace(",", ""), out _);
+        var isNumeric = long.TryParse(txtDebtAmount.Text.Replace(",", ""), out _);
         if (!isNumeric)
         {
             MessageBox.Show("لطفا مقدار بدهی را به درستی وارد کنید");
